% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billcalculate_function.R
\name{billcalculate}
\alias{billcalculate}
\title{takes an xml file of greenbutton format, gets the address and then estimates solar production}
\usage{
billcalculate(interval_data, cz, fuel, discount)
}
\arguments{
\item{the}{address of the house and the average annual consumption}
}
\value{
an array of average daily solar production
}
\description{
Takes energy consumption data, solar production data, climate zone, and
}
\author{
Christina Machak
}
