% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarbill_shift_function.R
\name{solarbill_shift}
\alias{solarbill_shift}
\title{takes a customers consumption data, estimated solar production and}
\usage{
solarbill_shift(interval_data, solar.prod, discount)
}
\arguments{
\item{interval_data, }{solar_prod, cz, fuel}
}
\value{
an array of average daily solar production
}
\description{
Takes energy consumption data, solar production data, climate zone, and
}
\author{
Christina Machak
}
