% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarproduction_function.R
\name{solarproduction}
\alias{solarproduction}
\title{takes an xml file of greenbutton format, gets the address and then estimates solar production}
\usage{
solarproduction(lat, lon, annual.consumption)
}
\arguments{
\item{the}{address of the house and the average annual consumption}
}
\value{
an array of average daily solar production
}
\description{
Takes greenbutton data,extracts hourly interval and returns daily mean consumption
}
\author{
Christina Machak
}
