% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarproductionAdv_function.R
\name{solarproductionAdv}
\alias{solarproductionAdv}
\title{takes an xml file of greenbutton format, gets the address and then estimates solar production}
\usage{
solarproductionAdv(lat, lon, system.size, tilt, azimuth)
}
\arguments{
\item{the}{address of the house and the average annual consumption}
}
\value{
an array of average daily solar production
}
\description{
An advanced version of the solar production calculator, allows the user to input tilt, azimuth and system size
}
\author{
Christina Machak
}
